requirements.txt
absl-py==2.0.0
array-record==0.5.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cycler==0.12.1
dm-tree==0.1.8
docker-pycreds==0.4.0
etils==1.5.2
flatbuffers==23.5.26
fsspec==2023.12.2
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.0
importlib-resources==6.1.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.4.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pathtools==0.1.2
Pillow==10.1.0
promise==2.3
protobuf==3.20.0
psutil==5.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
PyYAML==6.0.1
